requirements.txt
gradio==3.41.0
python-dotenv
pandas